pyproject.toml
[project]
name = "student-api"
version = "0.1.0"
description = "Student CRUD REST API with Django REST Framework"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14,<4.0",
    "djangorestframework-stubs>=3.16.7",
    "psycopg2-binary>=2.9,<3.0",
    "python-dotenv>=1.0,<2.0",
]
